Avoid buffer overflow with long filenames in write_auth_file()
Aligns buffer size for filenames with XauLockAuth in libXau

Closes: #7 ("Buffer overflow when constructing tmp_nam")
Reported-by: Mingjie Shen (@szsam)
Part-of: <https://gitlab.freedesktop.org/xorg/app/xauth/-/merge_requests/24>

+++ process.c
@@ -714,9 +714,14 @@ auth_initialize(const char *authfilename)
     FILE *authfp;
     Bool exists;
 
+    /*
+     * XauLockAuth in libXau limits file names to 1022 characters so it
+     * has room to append two characters in its 1025 character buffers.
+     */
     if (strlen(authfilename) > 1022) {
 	fprintf (stderr, "%s: authority file name \"%s\" too long\n",
 		 ProgramName, authfilename);
+	exit (1);
     }
     xauth_filename = authfilename;    /* used in cleanup, prevent race with
                                          signals */
@@ -822,16 +827,18 @@ auth_initialize(const char *authfilename)
 }
 
 static int
-write_auth_file(char *tmp_nam)
+write_auth_file(char *tmp_nam, size_t tmp_nam_size)
 {
     FILE *fp = NULL;
     int fd;
 
-    /*
-     * xdm and auth spec assumes auth file is 12 or fewer characters
-     */
-    strcpy (tmp_nam, xauth_filename);
-    strcat (tmp_nam, "-n");		/* for new */
+    /* Append "-n" for "new" */
+    int ret = snprintf(tmp_nam, tmp_nam_size, "%s-n", xauth_filename);
+    if (ret < 0 || ret >= tmp_nam_size) {
+        fprintf(stderr, "Error constructing filename: %s\n",
+                (ret < 0) ? "snprintf failed" : "buffer size is too small");
+        return -1;
+    }
     (void) unlink (tmp_nam);
     /* CPhipps 2000/02/12 - fix file unlink/fopen race */
     fd = open(tmp_nam, O_WRONLY | O_CREAT | O_EXCL, 0600);
@@ -907,7 +914,7 @@ auth_finalize(void)
 			"Writing", xauth_filename);
 	    }
 	    temp_name[0] = '\0';
-	    if (write_auth_file (temp_name) == -1) {
+	    if (write_auth_file (temp_name, sizeof(temp_name)) == -1) {
 		fprintf (stderr,
 			 "%s:  unable to write authority file %s\n",
 			 ProgramName, temp_name);
